navigation/requirements.txt
accelerate==0.28.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
backports-datetime-fromisoformat==2.0.3
boto3==1.34.69
botocore==1.34.162
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
datasets==2.16.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
docstring-parser==0.16
einops==0.6.1
einops-exts==0.0.4
environs==14.2.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.115.13
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gradio-client==0.2.9
h11==0.14.0
hf-transfer==0.1.9
hf-xet==1.1.5
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.0
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.7
latex2mathml==3.78.0
lmdb==1.6.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.9.4
mdurl==0.1.2
mmengine==0.10.7
mpmath==1.3.0
msgpack==1.1.1
msgpack-numpy==0.4.8
multidict==6.5.1
multiprocess==0.70.15
multiprocessing-logging==0.3.4
networkx==3.2.1
ninja==1.11.1.4
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.32.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
peft==0.5.0
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyarrow-hotfix==0.7
pydantic==1.10.8
pydantic-core==2.33.2
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.13.1
sentencepiece==0.1.99
sentry-sdk==2.31.0
setproctitle==1.3.6
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.20.3
tomli==2.2.1
tqdm==4.67.1
transformers==4.45.1
triton==2.1.0
typeguard==4.4.4
typing-extensions==4.14.0
typing-inspection==0.4.1
tyro==0.9.24
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
wandb==0.20.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==3.23.0
transformers==4.57.1
accelerate==1.11.0
deepspeed==0.18.2
ffmpeg-python==0.2.0
datasets==3.6.0
